# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Бэкенд Сервиса",
        "contact": {
            "name": "Docs developer",
            "url": "https://t.me/sigy922"
        },
        "version": "0.0.02"
    },
    "paths": {
        "/api/material": {
            "get": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Ручка",
                "responses": {
                    "200": {
                        "description": "http.StatusOK"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Добавление материала",
                "parameters": [
                    {
                        "description": "52",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Material"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "http.StatusOK"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Обновелние материала",
                "parameters": [
                    {
                        "description": "52",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Material"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "http.StatusOK"
                    }
                }
            }
        },
        "/api/material/csv": {
            "get": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Ручка",
                "responses": {
                    "200": {
                        "description": "http.StatusOK"
                    }
                }
            }
        },
        "/api/order": {
            "get": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Ручка",
                "responses": {
                    "200": {
                        "description": "http.StatusOK"
                    }
                }
            }
        },
        "/api/order/": {
            "post": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Добавление заявки",
                "parameters": [
                    {
                        "description": "52",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "http.StatusOK"
                    }
                }
            }
        },
        "/api/order/csv": {
            "get": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Ручка",
                "responses": {
                    "200": {
                        "description": "http.StatusOK"
                    }
                }
            }
        },
        "/api/order/status": {
            "patch": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Обновелние материала",
                "parameters": [
                    {
                        "description": "52",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "http.StatusOK"
                    }
                }
            }
        },
        "/api/product": {
            "get": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Ручка",
                "responses": {
                    "200": {
                        "description": "http.StatusOK"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Добавление продукта",
                "parameters": [
                    {
                        "description": "52",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "http.StatusOK"
                    }
                }
            }
        },
        "/api/product/csv": {
            "get": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Ручка",
                "responses": {
                    "200": {
                        "description": "http.StatusOK"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Авторизация"
                ],
                "summary": "Авторизация",
                "parameters": [
                    {
                        "description": "1",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AuthParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "http.StatusOK"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "security": [
                    {
                        "AuthToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Авторизация"
                ],
                "summary": "Регистрация",
                "parameters": [
                    {
                        "description": "1",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "http.StatusOK"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AuthParams": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.Material": {
            "type": "object",
            "required": [
                "count",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "count": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "reject_percent": {
                    "type": "number"
                }
            }
        },
        "models.MaterialSmallInfo": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "models.Order": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "materials": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MaterialSmallInfo"
                    }
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "reject_percent": {
                    "type": "number"
                }
            }
        },
        "models.RegisterParams": {
            "type": "object",
            "required": [
                "login",
                "name",
                "password",
                "surname"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "models.UpdateOrder": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AuthToken": {
            "description": "autorization token from auth_service",
            "type": "apiKey",
            "name": "sessionid",
            "in": "header"
        }
    }
}